www/requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.3
pymysql==1.0.2
mysql-connector-python
PyJWT

